% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fractal_app.r
\name{fractal_app}
\alias{fractal_app}
\title{fractal_app .}
\usage{
fractal_app()
}
\value{
a shiny app.
}
\description{
A shiny app to view fractals.
}
\examples{
\dontrun{
fractal_app()
}
}
\author{
Steven E. Pav \email{shabbychef@gmail.com}
}
\keyword{shiny}
