% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faroutman.r
\name{faroutman-NEWS}
\alias{faroutman-NEWS}
\title{News for package 'faroutman':}
\description{
News for package \sQuote{faroutman}

\newcommand{\CRANpkg}{\href{https://cran.r-project.org/package=#1}{\pkg{#1}}}
\newcommand{\faroutman}{\CRANpkg{faroutman}}
}
\section{\faroutman{} Initial Version 0.1.0 (2021-04-01) }{

\itemize{
\item first CRAN release.
}
}

